secure-file-transfer/pyproject.toml
[project]
name = "secure-file-transfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.119.1",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "jwt>=1.4.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.44",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pre-commit>=4.3.0",
    "ruff>=0.14.1",
]
